requirements.txt
google-api-python-client==2.88.0
supabase==1.0.3
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
tenacity==8.2.2